+++ x/vesting/keeper/gov.go
@@ -33,3 +33,34 @@ func (k Keeper) DeleteGovClawbackDisabled(ctx sdk.Context, addr sdk.AccAddress) 
 	key := append(types.KeyPrefixGovClawbackDisabledKey, addr.Bytes()...)
 	ctx.KVStore(k.storeKey).Delete(key)
 }
+
+// HasActiveClawbackProposal checks if there is an active clawback proposal for the given
+// vesting account address.
+func (k Keeper) HasActiveClawbackProposal(ctx sdk.Context, addr sdk.AccAddress) bool {
+	key := buildActiveAccountClawbackProposalKey(addr)
+
+	return ctx.KVStore(k.storeKey).Has(key)
+}
+
+// SetActiveClawbackProposal sets the given vesting account address as subject to an active governance clawback
+// proposal by writing it to store under the corresponding key.
+func (k Keeper) SetActiveClawbackProposal(ctx sdk.Context, addr sdk.AccAddress) {
+	key := buildActiveAccountClawbackProposalKey(addr)
+	ctx.KVStore(k.storeKey).Set(key, []byte{0x01})
+}
+
+// DeleteActiveClawbackProposal deletes the entry for the given vesting account address
+// from the store, indicating that there is no active governance clawback proposal for it.
+func (k Keeper) DeleteActiveClawbackProposal(ctx sdk.Context, addr sdk.AccAddress) {
+	key := buildActiveAccountClawbackProposalKey(addr)
+	ctx.KVStore(k.storeKey).Delete(key)
+}
+
+// buildActiveAccountClawbackProposalKey builds the key for the given account address prefixed with the governance clawback proposal key
+func buildActiveAccountClawbackProposalKey(addr sdk.AccAddress) []byte {
+	key := make([]byte, 0, len(types.KeyPrefixGovClawbackProposalKey)+len(addr.Bytes()))
+	key = append(key, types.KeyPrefixGovClawbackProposalKey...)
+	key = append(key, addr.Bytes()...)
+
+	return key
+}